update references for moving dbt-adapters into a package subdirectory
