+++ dbt-adapters/src/dbt/adapters/__version__.py
