+++ dbt-adapters/tests/unit/test_adapter_telemetry.py
@@ -1,4 +1,4 @@
-import dbt.adapters.__about__
+import dbt.adapters.__version__
 
 from dbt.adapters.base.impl import BaseAdapter
 from dbt.adapters.base.relation import AdapterTrackingRelationInfo
@@ -8,7 +8,7 @@ def test_telemetry_returns():
     res = BaseAdapter.get_adapter_run_info({})
 
     assert res.adapter_name == "base"
-    assert res.base_adapter_version == dbt.adapters.__about__.version
+    assert res.base_adapter_version == dbt.adapters.__version__.version
     assert res.adapter_version == ""
     assert res.model_adapter_details == {}
